requirements.txt
appdirs==1.4.3
dj-database-url==0.4.2
Django==1.11.1
django-cors-headers==2.0.2
django-material==0.13.1
djangorestframework==3.6.2
gunicorn==19.7.1
olefile==0.44
packaging==16.8
Pillow==4.1.1
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
Unipath==1.1
whitenoise==3.3.0
